Refactor project file management
