+++ EventPlanner/app/src/main/res/layout/fragment_homescreen.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".fragments.HomeScreenFragment">
+    tools:context=".fragments.home.HomeScreenFragment">
 
     <HorizontalScrollView
         android:layout_width="match_parent"
